Add deletion of movies

+++ src/models/movies.js
@@ -1,13 +1,7 @@
 const movieQuery = require("../../queries/movies");
 
-const createMovie = async ({ title, director, year, my_rating, image }) => {
-  let errors = [];
-  if (!title) errors.push("Title field is missing");
-  if (!director) errors.push("Director field is missing");
-  if (!year) errors.push("Year field is missing");
-  if (!my_rating) errors.push("My Rating field is missing");
-  if (!image) errors.push("image field is missing");
-
+const createMovie = async (movie) => {
+  const errors = checkFields(movie);
   if (errors.length > 0) {
     return {
       status: 400,
@@ -15,8 +9,8 @@ const createMovie = async ({ title, director, year, my_rating, image }) => {
       errors
     };
   }
-  const result = await movieQuery.createMovie({ title, director, year, my_rating, image })
 
+  const result = await movieQuery.createMovie(movie)
   return result;
 };
 
@@ -27,7 +21,7 @@ const getAllMovies = async () => {
 };
 
 const getMovieById = async (id) => {
-  if (typeof id !== "number") {
+  if (typeof Number(id) !== "number") {
     return {
       status: 404,
       message: "Not Found",
@@ -36,7 +30,6 @@ const getMovieById = async (id) => {
   }
 
   const movie = await movieQuery.getMovieById(id);
-
   if (!movie) {
     return {
       status: 404,
@@ -48,18 +41,41 @@ const getMovieById = async (id) => {
   return movie
 };
 
-const updateMovie = (id, body) => {
-  const result = movieQuery.updateMovie(id, body);
+const updateMovie = async (id, movie) => {
+  const movieById = await getMovieById(id);
+  if (movieById.status === 404) return movieById;
 
+  const errors = checkFields(movie);
+  if (errors.length > 0) {
+    return {
+      status: 400,
+      message: "Missing fields: ",
+      errors
+    };
+  }
 
+  const result = await movieQuery.updateMovie(id, movie);
+  return result;
 };
 
-const deleteMovie = (id) => {
-  const result = movieQuery.deleteMovie(id);
+const deleteMovie = async (id) => {
+  const movieById = await getMovieById(id);
+  if (movieById.status === 404) return movieById;
 
-
+  const result = await movieQuery.deleteMovie(id);
+  return result;
 };
 
+const checkFields = ({ title, director, year, my_rating, image }) => {
+  let errors = [];
+  if (!title) errors.push("Title field is missing");
+  if (!director) errors.push("Director field is missing");
+  if (!year) errors.push("Year field is missing");
+  if (!my_rating) errors.push("My Rating field is missing");
+  if (!image) errors.push("image field is missing");
+
+  return errors;
+};
 
 
 module.exports = {
